gitlab-ci: use custom Docker image and build in parallel with Clang and GCC
While Travis CI still seems to support only Ubuntu Precise or Trusty [1]
GitLab CI is much more flexible by using Docker (however, inside the
Travis VMs one seems to be able to make use of Docker which seems
cumbersome [2]).

A side effect of that limitation seems to be that almost all Travis runs
initially pull some dependencies (`apt install ...`) which can entirely
be skipped by providing a suitable Docker image to the GitLab registry.

Travis CI allows to configure a "build matrix" [3] which has a lot of
duplicated lines. The "anchor" feature of .gitlab-ci.yml allows to
inherit from templates. Both result in parallel builds.

While merging multiple templates works fine, merging multiple templates
with variables does not. In the following example, *clang overwrites all
variables set in the *debug template that is referenced before:

  .clang_template: &clang
      variables:
          COMPILER: clang++-6.0

  .gcc_template: &gcc
      variables:
          COMPILER: g++

  .debug_template: &debug
      variables:
          MODE: Debug

  .release_template: &release
      variables:
          MODE: Release

  clang-debug:
      <<: *build_definition
      <<: *debug                # MODE is unset/does not exist
      <<: *clang                # COMPILER is set

Build and push the image using the following commands

  docker build -t registry.gitlab.com/obruns/bluetoe/ubuntu-build:1.0 - < Dockerfile
  docker push registry.gitlab.com/obruns/bluetoe/ubuntu-build:1.0

Making use of stages [5] does not make sense in this context because
stages allow parallel execution of steps that result in artifacts that
are later combined. This requires that the stages deploy or publish the
artifacts for later consumption because each stage is a separate
container instance.

[1] https://docs.travis-ci.com/user/reference/overview/#Virtualisation-Environment-vs-Operating-System
[2] https://docs.travis-ci.com/user/reference/trusty
[3] https://docs.travis-ci.com/user/languages/cpp/#GCC-on-Linux
[4] https://docs.gitlab.com/ce/ci/yaml/README.html#anchors
[5] https://docs.gitlab.com/ee/ci/yaml/#stages