build scripts, separate dependencies (draft)
DO NOT MERGE

Attempt at a clear separation of dependencies between the different
submodules of bluetoe.

create a bluetoe::tests:test_tools library
create a bluetoe::utility library
create a bluetoe:sm (security manager) library

The dependencies are meant to be explicitly defined thru
target_link_libraries() in the CMakeLists.txt of each subfolder.

The tests would only need to link to the specific libraries they test.

The examples would link to a super_lib named "bluetoe::bluetoe" that has
all the internal dependencies in it.