source structure, add support for nrf52
DO NOT MERGE

This initial commit to enable easy building the examples for nrf51 or
nrf52 boards.

As the code in bindings/nrf51 is used for both the nrf51 and the nrf52,
for the momement there is no point in creating one bindings/nrf52 folder
because we would basically duplicate the code. Maybe the binding can be
renamed to bindings/nrfx

to build the examples for nrf52:
mkdir build_nrf52 && cd build_nrf52
cmake -DCMAKE_BUILD_TYPE=RelWithDebInfo
-DCMAKE_TOOLCHAIN_FILE=../cmake/gcc-arm-none-eabi.cmake -DSOC=nrf52
-DNRF5_SDK_ROOT=~/path/to/nrf5_sdk/ ..
make blinky.flash

to build for the nrf51, replace -DSOC=nrf52 with -DSOC=nrf51